Extract Witness types from shelley `Tx` module into new module `TxWits`
