+++ eras/shelley-ma/impl/src/Cardano/Ledger/Allegra/Translation.hs
@@ -24,7 +24,7 @@ import Cardano.Ledger.Shelley.API
 import qualified Cardano.Ledger.Shelley.LedgerState as LS
   ( returnRedeemAddrsToReserves,
   )
-import Cardano.Ledger.Shelley.Tx (decodeWits)
+import Cardano.Ledger.Shelley.TxWits (decodeWits)
 import Control.Monad.Except (throwError)
 import Data.Coerce (coerce)
 import qualified Data.Map.Strict as Map
