+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Examples/Updates.hs
@@ -39,12 +39,14 @@ import Cardano.Ledger.Shelley.PParams
   )
 import Cardano.Ledger.Shelley.Tx
   ( ShelleyTx (..),
-    WitnessSetHKD (..),
   )
 import Cardano.Ledger.Shelley.TxBody
   ( ShelleyTxBody (..),
     ShelleyTxOut (..),
     Wdrl (..),
+  )
+import Cardano.Ledger.Shelley.TxWits
+  ( WitnessSetHKD (..),
   )
 import Cardano.Ledger.Shelley.UTxO (UTxO (..), makeWitnessesVKey)
 import Cardano.Ledger.Slot
